+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Version 1.6.20
 
+### Bug fixes
+- Error callback when authentication failed [DE345468]
+    - Developer may need to update the App if using Proximity Login.
+    - com.ca.mas.core.auth.PollingRenderer.onAuthCodeReceived interface has been changed, the authorization code and state is provided.
+    - onAuthCodeReceived, developer should call com.ca.mas.foundation.MASUser.login(com.ca.mas.foundation.MASAuthCredentials, com.ca.mas.foundation.MASCallback<com.ca.mas.foundation.MASUser>) with com.ca.mas.foundation.MASAuthCredentialsAuthorizationCode
+    - Do not execute pending request when authentication failed. The pending queue will only be executed after authentication success.
+
 ### Deprecated Classes
 - MASOtpAuthFragment.java is removed, please use MASOtpActivity/MASOtpDialogFragment.
 - MASOtpSelectDeliveryChannelFragment.java is removed, please use MASOtpActivity/MASOtpDialogFragment.
@@ -15,7 +22,6 @@
 - com.ca.mas.connecta.client.MASConnectaClient.getTimeOutInMillis is removed, please use MASConnectOptions#getConnectionTimeout.
 - com.ca.mas.core.error.MAGError.getResultCode is removed.
 - com.ca.mas.core.error.MAGError.setResultCode is removed.
-
 
 # Version 1.6.10